requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.17.0
